# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "SkySwim Express API",
    "description": "SkySwim API used by the client",
    "contact": {
      "name": "Mostapha Rammo",
      "url": "skyswim.mrammo.ca",
      "email": "mr.rammo@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "localhost"
    },
    {
      "url": "https://skyswim.mrammo.ca",
      "description": "server"
    }
  ],
  "paths": {
    "/api/s3-signed-url": {
      "get": {
        "description": "Get and AWS.S3 presigned PUT url, allowing you to upload directly to the s3 bucket",
        "tags": [
          "/api"
        ],
        "parameters": [
          {
            "name": "contentType",
            "description": "Content-Type of the file to be uploaded using link",
            "in": "query",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "image/jpeg"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "url signed"
          },
          "400": {
            "description": "Invalid request. Content-Type must be provided"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "description": "Get list of 10 users based on matching search terms",
        "tags": [
          "/api"
        ],
        "parameters": [
          {
            "name": "text",
            "description": "Search for users matching this text",
            "in": "query",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "some example text"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "users returned"
          },
          "400": {
            "description": "Invalid request. Content-Type must be provided"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
	},
	"/api/post/user": {
		"get": {
			"description": "Get batch of user's posts based on page value",
			"tags": [
				"/api"
			],
			"parameters": [
				{
					"name": "page",
					"description": "page offset of posts",
					"in": "query",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"example": "1"
							}
						}
					}
				},
				{
					"name": "user",
					"description": "user id",
					"in": "query",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"example": "1fh289rg813h780e73801h"
							}
						}
					}
				}
			],
			"responses": {
				"200": {
					"description": "Posts returned"
				},
				"400": {
					"description": "Invalid request. Content-Type must be provided"
				},
				"default": {
					"description": "Unexpected error"
				}
			}
		}
	},
	"/api/post/feed": {
		"get": {
			"description": "Get batch of feed posts based on page value",
			"tags": [
				"/api"
			],
			"parameters": [
				{
					"name": "page",
					"description": "page offset of posts",
					"in": "query",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "string",
								"example": "1"
							}
						}
					}
				}
			],
			"responses": {
				"200": {
					"description": "Posts returned"
				},
				"400": {
					"description": "Invalid request. Content-Type must be provided"
				},
				"default": {
					"description": "Unexpected error"
				}
			}
		}
	},
    "/api/post": {
      "post": {
        "description": "Creates a post",
        "tags": [
          "/api"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "type": {
                    "type": "string",
                    "example": "image"
                  },
                  "blurb": {
                    "type": "string",
                    "example": "blurb"
                  },
                  "media": {
                    "type": "string",
                    "example": "https://fake.path.org/media"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Content posted"
          },
          "400": {
            "description": "Invalid request"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
	  }
	},
    "/api/follow": {
      "post": {
        "description": "Follow a user",
        "tags": [
          "/api"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "example": "3134hf1292ey793ehdq97e13"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User followed"
          },
          "400": {
            "description": "Invalid request"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    },
    "/api/unfollow": {
      "post": {
        "description": "Unfollow a user",
        "tags": [
          "/api"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user": {
                    "type": "string",
                    "example": "3134hf1292ey793ehdq97e13"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User unfollowed"
          },
          "400": {
            "description": "Invalid request"
          },
          "default": {
            "description": "Unexpected error"
          }
        }
      }
    }
  }
}